+++ CHANGELOG.md
@@ -13,6 +13,13 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- ids parameter to Transaction.query
+- ids parameter to Transfer.query
+
+### Fixed
+- BoletoPayment test case
+- UtilityPayment test case
 
 ## [0.9.0] - 2020-08-12
 ### Added